+++ src/test/fixtures/elements/ProductFixture.php
@@ -1,13 +1,20 @@
 <?php
+/**
+ * @link https://craftcms.com/
+ * @copyright Copyright (c) Pixel & Tonic, Inc.
+ * @license https://craftcms.github.io/license/
+ */
 
 namespace craft\commerce\test\fixtures\elements;
 
 use Craft;
+use craft\commerce\Plugin;
 use yii\base\ErrorException;
 use craft\base\Element;
 use craft\test\fixtures\elements\ElementFixture;
 use craft\commerce\elements\Product;
 use craft\commerce\elements\Variant;
+use yii\base\InvalidArgumentException;
 
 /**
  * Class ProductFixture.
@@ -17,10 +24,13 @@ use craft\commerce\elements\Variant;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @author Robuust digital | Bob Olde Hampsink <bob@robuust.digital>
  * @author Global Network Group | Giel Tettelaar <giel@yellowflash.net>
- * @since  3.2
+ * @since  2.1
  */
 class ProductFixture extends ElementFixture
 {
+    // Properties
+    // =========================================================================
+
     /**
      * {@inheritdoc}
      */
@@ -31,6 +41,9 @@ class ProductFixture extends ElementFixture
      */
     protected $productTypeIds = [];
 
+    // Public Methods
+    // =========================================================================
+
     /**
      * {@inheritdoc}
      */
@@ -38,17 +51,21 @@ class ProductFixture extends ElementFixture
     {
         parent::init();
 
-        /** @var \craft\commerce\Plugin */
-        $commerce = Craft::$app->getPlugins()->getPlugin('commerce');
-        /** @var \craft\commerce\services\ProductTypes */
-        $productTypesService = $commerce->getProductTypes();
+        // Ensure loaded
+        $commerce = Plugin::getInstance();
+        if (!$commerce) {
+            throw new InvalidArgumentException('Commerce plugin needs to be loaded before using the ProductFixture'.);
+        }
 
         // Get all product type id's
-        $productTypes = $productTypesService->getAllProductTypes();
+        $productTypes = $commerce->getProductTypes()->getAllProductTypes();
         foreach ($productTypes as $productType) {
             $this->productTypeIds[$productType->handle] = $productType->id;
         }
     }
+
+    // Protected Methods
+    // =========================================================================
 
     /**
      * {@inheritdoc}
@@ -56,37 +73,5 @@ class ProductFixture extends ElementFixture
     protected function isPrimaryKey(string $key): bool
     {
         return parent::isPrimaryKey($key) || in_array($key, ['typeId', 'title']);
-    }
-
-    /**
-     * Get element errors.
-     *
-     * @param Element $element
-     *
-     * @throws ErrorException
-     */
-    protected function getErrors(Element $element): void
-    {
-        $errors = $element->getErrorSummary(true);
-
-        foreach ($element->getVariants() as $variant) {
-            $errors = array_merge($errors, $variant->getErrorSummary(true));
-        }
-
-        throw new ErrorException(join(' ', array_filter($errors)));
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    protected function deleteElement(Element $element): void
-    {
-        $variants = Variant::find()->productId($element->id)->all();
-
-        foreach ($variants as $variant) {
-            parent::deleteElement($variant, true);
-        }
-
-        parent::deleteElement($element, true);
     }
 }